resolve some issues.

+++ common/src/main/java/com/nima/common/model/UrlIdSecond.kt
@@ -3,6 +3,6 @@ package com.nima.common.model
 data class UrlIdSecond(
     val ids: Long? = null,
     val regex: String? = null,
-    val startIndex: Int? = null,
-    val finishIndex: Int? = null,
+    val startIndex: Long? = null,
+    val finishIndex: Long? = null,
 )
